Improves BoxManager, adds new Boxes and implements a new function.
Improves BoxManager by storing the last box and adding the ability to jump to an box directly (will not slide, will fade).

Adds new Boxes about HTTP(S), Web Browser and TCP/IP, adds them to arrangement and fills the Technology Box with content.

Implements a new function, to be able to jump to a special box by location hash.